legacy/requirements.txt
albumentations==1.4.6
comet_ml==3.41.0
coremltools==7.2
crewai
common
easyocr==1.7.1
langchain_core==0.1.50
langchain_groq==0.1.3
llama-cpp-python
matplotlib==3.7.3
numpy
openai_whisper==20231117
opencv_python==4.9.0.80
pandas==2.0.3
Pillow==10.3.0
pyaudio
pydub==0.25.1
pygame==2.5.2
python-dotenv
pyttsx3==2.90
PyYAML==6.0.1
Requests
scipy
seaborn
SpeechRecognition==3.10.1
tensorflow
torch
torchaudio
torchvision
transformers
ultralytics==8.2.4
h5py
sentence-transformers
sounddevice
